requirements.txt
fastapi
uvicorn[standard]
python-dotenv
supabase
requests
jinja2
python-multipart
